fix: handleGoogleLogin 함수 util로 이동
해당 함수는 uril로 옮기기에 적합합니다
Landing 페이지 뿐아닌, Login이 필요한 로직에서 폭넓게 사용 가능합니다
